+++ main.go
@@ -14,7 +14,7 @@ import (
 )
 
 var (
-	version = "0.5.1"
+	version = "0.5.2"
 	commit  = "none"
 	date    = "unknown"
 )